Backend/requirements.txt
flask 
opencv-python
pymongo[srv]
Flask-pyMongo
python-dotenv
werkzeug
face-recognition
